requirements.txt
camelot-py[cv]==0.11.0
mysql-connector-python>=9.0.0
Flask>=3.0.0